#### Database Migration

<!-- If you update database schema, the migration number must be noted
here -->

NO

#### Description

Minor release catching main up with develop prior to the next upstream
sync.

#### Screenshot(s) (if UI related)

<details>
<!-- paste screenshots below here.-->

<!-- paste screenshots above here.  MAKE SURE THE ARE SFW -->
</details>

#### Todos

- [x] Tests
- [x] Translation Keys (./src/NzbDrone.Core/Localization/Core/en.json)

#### Issues Fixed or Closed by this PR

<!-- If multiple, put each on a separate line -->

- Fixes whisparr/whisparr-eros#XXXX

+++ src/Whisparr.Api.V3/Blocklist/BlocklistController.cs
@@ -8,6 +8,7 @@ using NzbDrone.Core.Datastore;
 using NzbDrone.Core.Indexers;
 using Whisparr.Http;
 using Whisparr.Http.Extensions;
+using Whisparr.Http.REST;
 using Whisparr.Http.REST.Attributes;
 
 namespace Whisparr.Api.V3.Blocklist
@@ -71,11 +72,14 @@ namespace Whisparr.Api.V3.Blocklist
         [HttpDelete("bulk")]
         [Consumes("application/json")]
         [Produces("application/json")]
-        public object Remove([FromBody] BlocklistBulkResource resource)
+        public void Remove([FromBody] BlocklistBulkResource resource)
         {
-            _blocklistService.Delete(resource.Ids);
+            if (resource.Ids == null || resource.Ids.Count == 0)
+            {
+                throw new BadRequestException("ids must be provided");
+            }
 
-            return new { };
+            _blocklistService.Delete(resource.Ids);
         }
     }
 }
